When determining trick winner, force players to match form of first player.
Previously, order of play was not taken into account. Thus, for example,
the first player may play A,K, and if second player follows with 3,3 in
the trick suit, the second player will incorrectly be the winner.

Now, we add a stage where we update the tractors played by later players
to be in the same form as the first player. In the example above, the 3,3
play would be treated as two tractors of rank=1,length=1.