Support flushes on first play in trick.
Only requirement on play is now that all cards should be in same
normalized suit. If a tractor in the flush is beaten by another
player, then the smallest beaten tractor is played, and all of the
cards in the original failed attempt are stored as failed_flush in
the state (and the player view).

UI is not updated yet, but can render failed_flush.